% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_mix.R
\name{hlpr_auto_constraints}
\alias{hlpr_auto_constraints}
\title{Auto-generate constraints from model return rate ranges}
\usage{
hlpr_auto_constraints(mrms, weekly_budget, bounds_multiplier = 3)
}
\description{
Auto-generate constraints from model return rate ranges
}
\keyword{internal}
