% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr_define_response_form.R
\name{hlpr_define_response_form}
\alias{hlpr_define_response_form}
\title{Define a response form for a nonlinear model}
\usage{
hlpr_define_response_form(type, x = NULL, y = NULL)
}
\arguments{
\item{type}{A character string specifying the type of response form.
Valid options are "logistic", "log_logistic", "gompertz", "reflected_gompertz", "weibull", and "reflected_weibull".}

\item{x}{A character string representing the name of the predictor variable.}

\item{y}{A character string representing the name of the response variable.}
}
\value{
A `brms` formula object representing the response form.
}
\description{
This function defines a response form for a nonlinear model based on the specified type.
It uses the `brms` package to create a formula for the response variable.
}
\details{
The function supports the following response forms:
\itemize{
  \item logistic: \eqn{y = c + ((d - c)/(1 + exp(b*(x - e))))}
  \item log_logistic: \eqn{y = c + ((d - c)/(1 + exp(b*(log(x) - log(e))))}
  \item gompertz: \eqn{y = c + (d - c) * exp(-exp(b * (x - e)))}
  \item reflected_gompertz: \eqn{y = c + (d - c) * (1 - exp(-exp(b * (-x + e))))}
  \item weibull: \eqn{y = c + (d - c) * exp(-exp(b * (log(x) - log(e))))}
  \item reflected_weibull: \eqn{y = c + (d - c) * (1 - exp(-exp(b * (-log(x) + log(e))))}
  }
}
