% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mrmfit.R
\name{hlpr_plot_diagnostics}
\alias{hlpr_plot_diagnostics}
\title{Plot brms diagnostics for an mrmfit object}
\usage{
hlpr_plot_diagnostics(mrm)
}
\arguments{
\item{mrm}{A fitted model object of class \code{mrmfit}.}
}
\value{
A patchwork plot object.
}
\description{
Produces trace plots and a posterior predictive check arranged vertically.
}
\keyword{internal}
