% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr_set_objective_function.R
\name{hlpr_set_objective_function}
\alias{hlpr_set_objective_function}
\title{Create an objective function for optimization}
\usage{
hlpr_set_objective_function(response_funs)
}
\arguments{
\item{response_funs}{A list of functions. Each function should take a single numeric input and return a numeric output.}

\item{prices}{An optional numeric vector of prices corresponding to each response function. If provided,}
}
\value{
A function that takes a numeric vector as input and returns the negative sum of the outputs from the response functions.
}
\description{
This function generates an objective function that can be used in optimization routines.
The objective function computes the negative sum of the outputs from a list of response functions
when evaluated at a given set of input parameters.
}
