% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrm_plot_response.R
\name{hlpr_unscale_data_points}
\alias{hlpr_unscale_data_points}
\title{Unscale observed data points for plotting}
\usage{
hlpr_unscale_data_points(mrm, x_var = "spend")
}
\arguments{
\item{mrm}{A fitted model object.}

\item{x_var}{Character; \code{"spend"} or \code{"units"}.}
}
\value{
A data frame with columns \code{x_plot_val} and \code{y_plot_val}.
}
\description{
Unscale observed data points for plotting
}
\keyword{internal}
