% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrm_params_table.R
\name{mrm_params_table}
\alias{mrm_params_table}
\title{Create a formatted table of model parameters}
\usage{
mrm_params_table(mrm, scaled = TRUE)
}
\arguments{
\item{mrm}{A model object.}

\item{scaled}{Logical indicating whether to return scaled parameters. Default is TRUE.}
}
\value{
A gt table of model parameters.
}
\description{
Create a formatted table of model parameters
}
