% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrm_returns_ranges.R
\name{mrm_returns_ranges}
\alias{mrm_returns_ranges}
\title{Get Ranges of Maximum Marginal Returns and Average Returns
This function computes the ranges of x values corresponding to the maximum marginal returns (MR)
and average returns (AR) for a list of marginal response models (MRMs).}
\usage{
mrm_returns_ranges(
  mrm,
  xrange = NULL,
  length.out = 1000,
  scaled = TRUE,
  cost_per_unit = 1,
  response_rate = 1
)
}
\arguments{
\item{mrm}{A list of fitted model objects (e.g., brmsfit objects).}

\item{xrange}{A numeric vector of length 2 specifying the range of x values to consider. If NULL, the range is determined from the data.}

\item{length.out}{An integer specifying the number of points to generate for the x-axis. Default is 1000.}

\item{scaled}{A logical value indicating whether to use scaled values for the calculations. Default is TRUE.}
}
\value{
A data frame containing the channels and their corresponding x ranges for maximum MR and AR.
}
\description{
Get Ranges of Maximum Marginal Returns and Average Returns
This function computes the ranges of x values corresponding to the maximum marginal returns (MR)
and average returns (AR) for a list of marginal response models (MRMs).
}
\details{
The function computes the MR and AR for each model in the list and identifies the x values where MR and AR are maximized. It returns a data frame with the channel names and their respective x ranges.
}
