% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_optimal_mix.R
\name{plot_optimal_mix}
\alias{plot_optimal_mix}
\title{Plot Optimal Mix}
\usage{
plot_optimal_mix(optimal_mix)
}
\arguments{
\item{response_funs}{A named list of response functions.}

\item{res}{The result object from the optimization containing the optimal solution.}
}
\value{
A ggplot object showing the response curves and optimal points.
}
\description{
This function plots the response curves along with the optimal points determined by the optimization results.
}
