% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mrm_prior.R
\name{print.mrm_prior}
\alias{print.mrm_prior}
\title{Print method for mrm_prior objects}
\usage{
\method{print}{mrm_prior}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mrm_prior}, created by [mrmopt_prior()].}

\item{...}{Currently unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print method for mrm_prior objects
}
