% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mrmfit_list.R
\name{print.mrmfit_hier_unit}
\alias{print.mrmfit_hier_unit}
\title{Print method for hierarchical unit views}
\usage{
\method{print}{mrmfit_hier_unit}(x, ...)
}
\arguments{
\item{x}{A \code{mrmfit_hier_unit} object from \code{\link{as_mrmfit_list}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for hierarchical unit views
}
