% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_models_comparison.R
\name{response_models_comparison}
\alias{response_models_comparison}
\title{Compare Response Curve Models}
\usage{
response_models_comparison(
  x = seq(0, 1, by = 0.01),
  b = -8,
  c = 0,
  d = 10,
  e = 0.5
)
}
\arguments{
\item{x}{A numeric vector representing the input values for the response curves. Default is seq(0, 1, by = 0.01).}

\item{b}{A numeric value representing the 'b' parameter for the response curves. Default is -8.}

\item{c}{A numeric value representing the 'c' parameter for the response curves. Default is 0.}

\item{d}{A numeric value representing the 'd' parameter for the response curves. Default is 10.}

\item{e}{A numeric value representing the 'e' parameter for the response curves. Default is 0.5.}
}
\description{
This function generates a comparison plot of various response curve models using specified parameters.
}
