% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_response.R
\name{fit_response}
\alias{fit_response}
\title{Fit a response curve model using brms}
\usage{
fit_response(data, x = NULL, y = NULL, auto = TRUE, type = "gompertz", ...)
}
\arguments{
\item{data}{A data frame containing the data to be fitted.}

\item{x}{The name of the independent variable (predictor).}

\item{y}{The name of the dependent variable (response).}

\item{auto}{Logical indicating whether to automatically scale the data and set priors. Default is TRUE.}

\item{type}{The type of response curve model to fit. Options are "gompertz", "logistic", "weibull", or "exponential".}

\item{...}{Additional arguments to be passed to the brms::brm function.}
}
\value{
A fitted model object.
}
\description{
This function fits a response curve model using the brms package.
}
\details{
The function fits a response curve model using the specified type and returns the fitted model object.
}
