% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr_default_prior_for_scaled_data.R
\name{hlpr_default_prior_for_scaled_data}
\alias{hlpr_default_prior_for_scaled_data}
\title{Helper function to set default priors for scaled data}
\usage{
hlpr_default_prior_for_scaled_data(scaled_data, x, y, scale_method)
}
\arguments{
\item{scale_method}{The method used for scaling the data. Currently, only "min_max" is supported.}
}
\value{
A list of priors for the parameters of the four-parameter logistic model.
}
\description{
This function sets default priors for the parameters of the four-parameter logistic model when the data has been scaled. The priors are based on the expected range of the parameters after scaling.
}
