% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr_get_weekly_spend.R
\name{hlpr_get_weekly_spend}
\alias{hlpr_get_weekly_spend}
\title{Get Weekly Spend helper function}
\usage{
hlpr_get_weekly_spend(mrm)
}
\arguments{
\item{mrm}{An MRM object containing the scaling method, scaling values, and data.}
}
\value{
A numeric value representing the weekly spend.
}
\description{
This function calculates the weekly spend based on the scaling method and values provided in the MRM object.
}
\details{
The function checks the scaling method used in the MRM object and calculates the weekly spend accordingly. If the scaling method is "min_max", it uses the mean of the data and the min-max scaling values to calculate the weekly spend. If the scaling method is "std", it uses the mean of the data and the standard deviation scaling values to calculate the weekly spend. If an invalid scaling method is provided, an error is raised.
}
