% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr_scale_data.R
\name{hlpr_scale_data}
\alias{hlpr_scale_data}
\title{Helper function to scale the data for model prep}
\usage{
hlpr_scale_data(data, x, y, scale_method)
}
\arguments{
\item{data}{The input data frame containing the x and y variables to be scaled.}
}
\value{
A list containing the scaled data, the inferred x-range, and the min-max values used for scaling.
}
\description{
This function scales the data using min-max scaling and also calculates the inferred x-range for the model. It returns a list containing the scaled data, the inferred x-range, and the min-max values used for scaling.
}
