% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrm_infer.R
\name{mrm_infer}
\alias{mrm_infer}
\title{Infer response from a fitted model}
\usage{
mrm_infer(mrm, xrange = NULL, length.out = 1000, scaled = TRUE)
}
\arguments{
\item{mrm}{A fitted model object.}

\item{xrange}{A numeric vector of length 2 specifying the range of x values for prediction. Default is NULL, which uses the range of x in the data.}

\item{length.out}{An integer specifying the number of points to predict. Default is 1000.}

\item{scaled}{A logical indicating whether the model was fitted on scaled data. Default is TRUE.}
}
\value{
A data frame containing the predicted response values and the model response.
}
\description{
This function infers the response from a fitted model object.
}
\details{
The function infers the response from the fitted model object and returns a data frame with the predicted response values and the model response.
}
