% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrm_params.R
\name{mrm_params}
\alias{mrm_params}
\title{Extract parameters from a fitted model}
\usage{
mrm_params(rc_fit, scaled = TRUE)
}
\arguments{
\item{rc_fit}{A fitted model object.}

\item{scaled}{A logical indicating whether the model was fitted on scaled data. Default is TRUE.}
}
\value{
A list containing the center, lower, and upper bounds of the parameters.
}
\description{
This function extracts the parameters from a fitted model object.
}
\details{
The function extracts the parameters from the fitted model object and returns them in a list.
}
