% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrm_plot_response.R
\name{mrm_plot_response}
\alias{mrm_plot_response}
\title{Plot the response of a fitted model}
\usage{
mrm_plot_response(
  mrm,
  xrange = NULL,
  length.out = 1000,
  points = TRUE,
  scaled = TRUE,
  markup = FALSE
)
}
\arguments{
\item{mrm}{A fitted model object. It can be a brmsfit object or a list of brmsfit objects.}

\item{xrange}{A vector of length 2 specifying the range of x values to plot. If NULL, the range of x values in the data is used.}

\item{length.out}{An integer specifying the number of points to generate for the x-axis. Default is 1000.}

\item{points}{A logical value indicating whether to plot the data points. Default is TRUE.}

\item{scaled}{A logical value indicating whether to plot the scaled response. Default is TRUE.}

\item{markup}{A logical value indicating whether to add markup lines to the plot. Default is FALSE.}
}
\value{
A ggplot object.
}
\description{
This function plots the response of a fitted model.
}
\details{
The function plots the response of a fitted model object. It uses ggplot2 to create the plot and includes a title and subtitle with information about the model.
}
