% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrm_plot_return.R
\name{mrm_plot_return}
\alias{mrm_plot_return}
\title{Plot Marginal and Absolute Rates of Return for Multiple Resource Models}
\usage{
mrm_plot_return(mrm, xrange = c(0, 2), ncol = 2, location = "center")
}
\arguments{
\item{mrm}{A list of fitted model objects (e.g., brmsfit objects).}

\item{xrange}{A numeric vector of length 2 specifying the range of x values to consider. If NULL, the range is determined from the data.}

\item{ncol}{An integer specifying the number of columns in the facet wrap. Default is 2.}

\item{location}{A character string specifying which location to use for MR and AR calculations. Options are "center", "lower", or "upper". Default is "center".}
}
\value{
A ggplot object visualizing the MR and AR for each model.
}
\description{
This function generates a plot comparing the Marginal Rate of Return (MR) and Absolute Rate of Return (AR) for multiple resource models. It highlights key points such as the maximum MR and the intersection point where MR equals AR.
}
\details{
The function computes the MR and AR for each model in the list and creates a faceted plot. It highlights the maximum MR point in red and the intersection point where MR equals AR in blue. A shaded green area indicates the range between these two points.
}
