% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmopt-package.R
\docType{package}
\name{mrmopt-package}
\alias{mrmopt}
\alias{mrmopt-package}
\title{mrmopt: mrmopt}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Package for media response modeling and optimization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bdshaff.github.io/mrmopt/}
}

}
\author{
\strong{Maintainer}: Ben Denis Shaffer \email{my.blogdown@gmail.com}

}
\keyword{internal}
